requirements.txt
jira==3.8.0
python-dotenv==1.1.0 